move traits to lib.rs
